Refactor synergy calculation
Move synergy calculation to 'SynergyCalculator classes, modeled using
ISynergyCalculator.

Add a SimpleSynergyCalculator class and move synergy related fields and
methods from DIMArmourPieceFactory to this new class.

Add a StaticSynergyCalculator class that simple returns a string passed
to the constructor.

Amend ArmourPiece to take an ISynergyCalculator and use that to implement
the getter for the Synergy property.

Remove the setter for the Synergy property.

Amend ArmourPieceFactory to create a StaticSynergyCalculator for each
ArmourPiece.

Amend DIMArmourPieceFactory to create a SimpleSynergyCalculator for each
ArmourPiece.

Add an Interfaces folder and namespace to hold both the new ISynergyCalculator
interface and the existing IArmourPieceFactory interface.