Add pick one of each code.
Add a function that picks one of each legendary piece, choosing the one
with the highest synergy, before processing the rest of the list. This
results in a more aggressive delete list, but keeps more distint pieces
with more synergy.

Also add a Tier property to ArmourPiece and use that are part of
determining whether to add perks to the found perks list.